refactor: 型定義ファイルをモデル毎に分割 #183
保守性向上のため
