+++ frontend/src/components/models/task/ChildrenTasksDetails.tsx
@@ -9,7 +9,7 @@ import {
   Typography,
 } from "@mui/material";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
-import { ChildTask } from "../../../types";
+import { ChildTask } from "../../../types/taskTypes";
 import { TaskContentTypography } from "./TaskContentTypography";
 import { DatetimeFormat } from "../../ui/DatetimeFormat";
 import { PriorityStack } from "./PriorityStack";
