+++ frontend/src/components/pages/DivisionsNew.tsx
@@ -16,12 +16,12 @@ import {
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
 import { axiosInstance } from "../../utils/axios";
+import { User } from "../../types/userTypes";
+import { DivisionTask } from "../../types/taskTypes";
 import {
-  DivisionTask,
   DivisionsCreateResponse,
   DivisionsNewResponse,
-  User,
-} from "../../types";
+} from "../../types/divisionTypes";
 import { AuthContext } from "../../providers/AuthProvider";
 import { SnackbarContext } from "../../providers/SnackbarProvider";
 import { TasksForm } from "../models/task/TasksForm";
