+++ frontend/src/components/pages/SignUp.tsx
@@ -5,7 +5,7 @@ import { Button, Stack, TextField, Typography } from "@mui/material";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
 import { AxiosRequestConfig, AxiosResponse } from "axios";
 import { axiosInstance } from "../../utils/axios";
-import { AuthResponse, PasswordState } from "../../types/index";
+import { AuthResponse, PasswordState } from "../../types/userTypes";
 import { AuthContext } from "../../providers/AuthProvider";
 import { SnackbarContext } from "../../providers/SnackbarProvider";
 import { PasswordTextfield } from "../models/user/PasswordTextfield";
